+++ example/lib/tabs/yt_web.dart
@@ -43,6 +43,9 @@ class YTWeb extends StatelessWidget {
             ),
             customPlayerWidgets: CustomPlayerWidgets().copyWith(
               thumbnailFit: BoxFit.fitWidth,
+              loadingWidget: Center(
+                child: CircularProgressIndicator(color: Colors.white),
+              ),
             ),
             playerTheme: OmniVideoPlayerThemeData().copyWith(
               icons: VideoPlayerIconTheme().copyWith(error: Icons.warning),